Replace fire-mode module with browser-mode and move relevant components there
