+++ app/build.gradle
@@ -420,8 +420,8 @@ dependencies {
     implementation project(':fingerprint-protection-impl')
     implementation project(':fingerprint-protection-store')
 
-    implementation project(':fire-mode-api')
-    implementation project(':fire-mode-impl')
+    implementation project(':browser-mode-api')
+    implementation project(':browser-mode-impl')
 
     implementation project(':element-hiding-impl')
     implementation project(':element-hiding-store')
