--- app/src/main/java/com/duckduckgo/app/browser/mode/RealBrowserModeStateHolder.kt
@@ -1,39 +1,0 @@
-/*
- * Copyright (c) 2026 DuckDuckGo
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *     http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-
-package com.duckduckgo.app.browser.mode
-
-import com.duckduckgo.browser.api.mode.BrowserMode
-import com.duckduckgo.browser.api.mode.BrowserModeStateHolder
-import com.duckduckgo.di.scopes.AppScope
-import com.squareup.anvil.annotations.ContributesBinding
-import dagger.SingleInstanceIn
-import kotlinx.coroutines.flow.MutableStateFlow
-import kotlinx.coroutines.flow.StateFlow
-import kotlinx.coroutines.flow.asStateFlow
-import kotlinx.coroutines.flow.update
-import javax.inject.Inject
-
-@SingleInstanceIn(AppScope::class)
-@ContributesBinding(AppScope::class)
-class RealBrowserModeStateHolder @Inject constructor() : BrowserModeStateHolder {
-    private val _currentMode = MutableStateFlow(BrowserMode.REGULAR)
-    override val currentMode: StateFlow<BrowserMode> = _currentMode.asStateFlow()
-
-    override fun switchTo(mode: BrowserMode) {
-        _currentMode.update { mode }
-    }
-}
