+++ browser-mode/browser-mode-impl/build.gradle
@@ -1,0 +1,60 @@
+/*
+ * Copyright (c) 2026 DuckDuckGo
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+plugins {
+    id 'com.android.library'
+    id 'kotlin-android'
+    id 'com.squareup.anvil'
+    id 'com.google.devtools.ksp'
+}
+
+apply from: "$rootProject.projectDir/gradle/android-library.gradle"
+
+android {
+    namespace 'com.duckduckgo.browsermode.impl'
+    anvil {
+        generateDaggerFactories = true
+    }
+    testOptions {
+        unitTests {
+            includeAndroidResources = true
+        }
+    }
+}
+
+dependencies {
+    ksp project(':anvil-ksp')
+    implementation project(path: ':anvil-annotations')
+
+    implementation project(path: ':browser-api')
+    implementation project(path: ':browser-mode-api')
+    implementation project(path: ':common-utils')
+    implementation project(path: ':di')
+    implementation project(path: ':feature-toggles-api')
+
+    implementation Google.dagger
+    implementation KotlinX.coroutines.core
+    implementation AndroidX.webkit
+    implementation AndroidX.dataStore.preferences
+
+    testImplementation Testing.junit4
+    testImplementation KotlinX.coroutines.test
+    testImplementation "org.mockito.kotlin:mockito-kotlin:_"
+    testImplementation Testing.robolectric
+    testImplementation AndroidX.test.ext.junit
+    testImplementation project(':common-test')
+    testImplementation project(':feature-toggles-test')
+}
