+++ build.gradle
@@ -177,6 +177,7 @@ subprojects {
                                 && dependencyPath != ":navigation-api"
                                 && dependencyPath != ":js-messaging-api"
                                 && dependencyPath != ":browser-api"
+                                && dependencyPath != ":browser-mode-api"
                         ) {
                             if (projectPath.endsWith(":feature-toggles-api") &&
                                     (dependencyPath == ":experiments-api"
