+++ duckchat/duckchat-impl/build.gradle
@@ -32,6 +32,7 @@ dependencies {
     implementation project(':design-system')
     implementation project(':common-utils')
     implementation project(':browser-api')
+    implementation project(':browser-mode-api')
     implementation project(':browser-ui')
     implementation project(':js-messaging-api')
     implementation project(':downloads-api')
